+++ changelog.md
@@ -9,6 +9,7 @@ _(most recent changes are listed on top):_
 - `D` added for "remove all subtasks".
 - `d` added for "duplicate instant" (DuplicateMod needed).
 - `C` added for "edit last comment".
+- `y` added for "scroll to comment section".
 
 
 ## v2.2